requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
pydantic==2.8.2
pydantic-settings==2.4.0
httpx==0.27.0
python-dotenv==1.0.1
sqlalchemy==2.0.32
oracledb==2.4.1
